requirements.txt
alembic==1.13.3
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
Mako==1.3.5
MarkupSafe==3.0.1
marshmallow==3.23.0
packaging==24.1
passlib==1.7.4
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
